nlp_pipeline/pipeline_requirements.txt
# Core data
pandas>=2.0
numpy>=1.24
polars>=0.20  # optional fast alternative

# NLP / ML
scikit-learn>=1.3
sentence-transformers>=2.2
transformers>=4.35
torch>=2.1

# Text processing
regex>=2023.0
emoji>=2.8
langdetect>=1.0.9
pydantic>=2.0

# Config
pyyaml>=6.0

# LLM adjudication (optional)
anthropic>=0.25

# Visualization / notebooks
matplotlib>=3.7
seaborn>=0.13
jupyter>=1.0
ipykernel>=6.0

# Testing
pytest>=7.0
pytest-cov>=4.0

# Utilities
python-dotenv>=1.0
tqdm>=4.60
